Add work-around for using CheckLatticeNodeOperations
Related to https://github.com/simphony/simphony-common/issues/216
For some tests in CheckLatticeNodeOperations it is expected that
the `container_factory` has produced a lattice with nodes
that are empty.  And that then these data of the lattice nodes can
be updated with cuba provided by the 'supported_cuba'. This doesn't
match the behaviour of JYULatticeProxy and it requires that the
supported cuba appear in the `external_node_data` parameter at object
initialization.  This has the effect that what CUBA appear on each
node.data cannot be changed later.